+++ openmdao/devtools/debug.py
@@ -3,7 +3,6 @@
 from __future__ import print_function
 
 import sys
-import os
 
 from resource import getrusage, RUSAGE_SELF, RUSAGE_CHILDREN
