+++ openmdao/solvers/linear/linear_block_gs.py
@@ -2,7 +2,7 @@
 
 from six.moves import range
 
-from openmdao.solvers.ln_solver import BlockLinearSolver
+from openmdao.solvers.solver import BlockLinearSolver
 
 
 class LinearBlockGS(BlockLinearSolver):
