+++ openmdao/solvers/linear/petsc_ksp.py
@@ -11,7 +11,7 @@ try:
 except ImportError:
     PETSc = None
 
-from openmdao.solvers.ln_solver import LinearSolver
+from openmdao.solvers.solver import LinearSolver
 from openmdao.utils.general_utils import warn_deprecation
