+++ openmdao/solvers/nonlinear/nonlinear_block_jac.py
@@ -1,5 +1,5 @@
 """Define the NonlinearBlockJac class."""
-from openmdao.solvers.nl_solver import NonlinearSolver
+from openmdao.solvers.solver import NonlinearSolver
 
 
 class NonlinearBlockJac(NonlinearSolver):
