+++ openmdao/vectors/petsc_vector.py
@@ -4,6 +4,7 @@ import numpy as np
 from petsc4py import PETSc
 
 from six import iteritems, itervalues
+from six.moves import range
 
 from openmdao.vectors.default_vector import DefaultVector, DefaultTransfer
 from openmdao.utils.mpi import MPI
@@ -53,9 +54,6 @@ class PETScTransfer(DefaultTransfer):
                 transfer = PETSc.Scatter().create(out_petsc, out_indexset, in_petsc, in_indexset)
 
                 transfers[key] = transfer
-
-        del self._out_inds
-        del self._in_inds
 
     def __call__(self, in_vec, out_vec, mode='fwd'):
         """
